Adds shared Dialog enum
