+++ components/form-builder/app/responses/Dialogs/DialogStates.ts
@@ -1,0 +1,12 @@
+export enum DialogStates {
+  EDITTING,
+  SENDING,
+  SENT,
+  MIN_ERROR,
+  MAX_ERROR,
+  FORMAT_ERROR,
+  FAILED_ERROR,
+  UNKNOWN_ERROR,
+  // Note: only used in ReportDialog
+  DESCRIPTION_EMPTY_ERROR,
+}
